requirements.txt
# FastAPI e servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Processamento de imagens
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.26.3

# YOLO
ultralytics==8.1.0
torch>=2.1.0
torchvision>=0.16.0

# OCR
easyocr==1.7.1
pytesseract==0.3.10

# Utilidades
python-dotenv==1.0.0
